Create jlstm.py
The file is not meant to work as it is. You need to have the input data, which I cannot put here because of copyright issues.

I used whole market data, and hence the need to input the specific stock code.  However, data can be easily obtained from Yahoo.co.jp individual stock data and adjust the above for individual stock data.